Split TreeListenerWrapper into two java files
A java file should only expose a single public class.
